requirements.txt
Django
django-extensions
djangorestframework
mysqlclient
